[use common header & lib from rust crates]
