+++ DashSync.podspec
@@ -32,8 +32,8 @@ Pod::Spec.new do |s|
   s.resource_bundles = {'DashSync' => ['DashSync/shared/*.xcdatamodeld', 'DashSync/shared/MappingModels/*.xcmappingmodel', 'DashSync/shared/*.plist', 'DashSync/shared/*.lproj', 'DashSync/shared/MasternodeLists/*.dat', 'DashSync/shared/*.json']}
   
   s.framework = 'Foundation', 'SystemConfiguration', 'CoreData', 'BackgroundTasks', 'Security'
-  s.ios.vendored_libraries = 'DashSync/lib/libmerkiOS.a', 'DashSync/lib/libmndiff_ios.a'
-  s.macos.vendored_libraries = 'DashSync/lib/libmerkMacOS.a', 'DashSync/lib/libmndiff_macos.a'
+  s.ios.vendored_libraries = 'DashSync/lib/*_ios.a'
+  s.macos.vendored_libraries = 'DashSync/lib/*_macos.a'
   s.ios.framework = 'UIKit'
   s.macos.framework = 'Cocoa'
   s.compiler_flags = '-Wno-comma'
