--- DashSync/lib/libmerkMacOS.a
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9b7f5f1e624d2c97b7af9447ad7b25df9123011892f472b8417c042710a3ca05
-size 35004232
