--- DashSync/lib/libmerkiOS.a
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:548cbdeac2a84a82c5c35b3d7edd7b95f8c168f83a9539cf54510df940f5ae19
-size 39369328
