--- DashSync/lib/libmndiff_ios.a
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d23cc46651b449dc6b1de6b96b09bdb52f69555ef004422058c92e330aaee7a0
-size 38289792
