--- DashSync/lib/libmndiff_macos.a
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7dff094f7e0482003a8b4ab77b964bf43e45e70b274471d1525ebf135c507f2b
-size 34242720
