+++ DashSync/shared/Models/DAPI/DSPlatformTreeQuery.h
@@ -15,7 +15,7 @@
 //  limitations under the License.
 //
 
-#import "merk.h"
+#import "dash_shared_core.h"
 #import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN
