+++ DashSync/shared/Models/Masternode/DSSimplifiedMasternodeEntry+Mndiff.h
@@ -17,7 +17,7 @@
 
 #import "DSChain.h"
 #import "DSSimplifiedMasternodeEntry.h"
-#import "mndiff.h"
+#import "dash_shared_core.h"
 #import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN
